refactor(FOROME-1187): general refactoring
