+++ src/components/conditions/components/approx-state.tsx
