+++ src/components/conditions/compound-request/compound-request-condition.tsx
@@ -4,21 +4,23 @@ import { observer } from 'mobx-react-lite'
 import { ApproxNameTypes } from '@core/enum/approxNameTypes'
 import { FuncStepTypesEnum } from '@core/enum/func-step-types-enum'
 import { ModeTypes } from '@core/enum/mode-types-enum'
-import { AprroxAndState } from '@components/conditions/compound-request/components/approx-state'
-import { InheritanceModeSelect } from '@pages/filter/dtree/components/query-builder/components/inheritance-mode-select'
+import { AprroxAndState } from '@components/conditions/components/approx-state'
+import { InheritanceModeSelect } from '@components/conditions/components/inheritance-mode-select'
 import { AllNotMods } from '@pages/filter/dtree/components/query-builder/ui/all-not-mods'
 import { DividerHorizontal } from '@pages/filter/refiner/components/middle-column/components/divider-horizontal'
 import { TRequestCondition } from '@service-providers/common'
-import { getFilteredRequestCondition } from '@utils/function-panel/getFilteredRequestCondition'
-import { getApproxValue } from '@utils/getApproxValue'
-import { getResetType } from '@utils/getResetType'
 import { DisabledVariants } from '../components/disabled-variants'
-import { ControlButtons } from './components/control-buttons'
+import {
+  getApproxValue,
+  getPreparedScenario,
+  getScenarioName,
+} from '../utils/conditions.utils'
 import { RequestConditions } from './components/request-conditions'
+import { RequestsAmountControlButtons } from './components/requests-amount-control-buttons'
 import { ICompoundRequestProps } from './compound-request.interface'
 import {
+  getFilteredRequestCondition,
   getNewRequestCondition,
-  getPreparedScenario,
 } from './compound-request.utils'
 
 export const CompoundRequestCondition = observer(
@@ -43,7 +45,7 @@ export const CompoundRequestCondition = observer(
 
     const [approx, setApprox] = useState<ApproxNameTypes>(initialApprox)
 
-    const [preparedValue, setPreparedValue] = useState<string>('')
+    const [preparedScenarioName, setPreparedScenarioName] = useState<string>('')
 
     const [activeRequestIndex, setActiveRequestIndex] = useState<number>(
       requestCondition.length - 1,
@@ -51,14 +53,16 @@ export const CompoundRequestCondition = observer(
 
     const toggleMode = (mode: ModeTypes) => {
       setMode(currentMode => (currentMode === mode ? undefined : mode))
-
       onTouch?.()
     }
 
-    const handleSetPreparedValue = (preparedValue: string) => {
-      setPreparedValue(preparedValue)
+    const handleSetPreparedScenario = (preparedScenarioName: string) => {
+      setPreparedScenarioName(preparedScenarioName)
 
-      const preparedScenario = getPreparedScenario(preparedValue, problemGroups)
+      const preparedScenario = getPreparedScenario(
+        preparedScenarioName,
+        problemGroups,
+      )
 
       const clonedRequestCondition = [...requestCondition]
       clonedRequestCondition[activeRequestIndex][1] = preparedScenario
@@ -83,38 +87,41 @@ export const CompoundRequestCondition = observer(
       )
 
       setRequestCondition(newRequestCondition)
-
-      setPreparedValue('')
-
+      setPreparedScenarioName('')
       onTouch?.()
     }
 
     const handleClearRequestCondition = () => {
-      setPreparedValue('')
-      setRequestCondition([])
+      setPreparedScenarioName('')
+      setRequestCondition([[1, {}]])
+      setMode(undefined)
     }
 
-    const handleActiveRequest = (requestBlockIndex: number): void => {
+    const handleSetActiveRequestCondition = (
+      requestBlockIndex: number,
+    ): void => {
       setActiveRequestIndex(requestBlockIndex)
 
       const currentRequest = requestCondition[requestBlockIndex]
 
-      setPreparedValue(getResetType(currentRequest[1]) || '')
+      setPreparedScenarioName(getScenarioName(currentRequest[1]) || '')
+      onTouch?.()
     }
 
-    const handleRequestConditionNumber = (
+    const onChangeRequestConditionNumber = (
       requestBlockIndex: number,
       value: number,
     ): void => {
       if (value < 0) return
 
       const newRequestCondition = [...requestCondition]
-
       newRequestCondition[requestBlockIndex][0] = value
+
       setRequestCondition(newRequestCondition)
+      onTouch?.()
     }
 
-    const handleRequestBlocksAmount = (
+    const handleChangeRequestsAmount = (
       type: string,
       activeRequestIndex: number,
     ): void => {
@@ -123,19 +130,19 @@ export const CompoundRequestCondition = observer(
         const newRequestCondition = [...requestCondition, emptyBlock]
 
         setActiveRequestIndex(newRequestCondition.length - 1)
-
         setRequestCondition(newRequestCondition)
-        setPreparedValue('')
+        setPreparedScenarioName('')
       } else {
         const newRequestCondition = requestCondition.filter(
           (_, index) => index !== activeRequestIndex,
         )
 
         setActiveRequestIndex(newRequestCondition.length - 1)
-
         setRequestCondition(newRequestCondition)
-        setPreparedValue(
-          getResetType(newRequestCondition[newRequestCondition.length - 1][1]),
+        setPreparedScenarioName(
+          getScenarioName(
+            newRequestCondition[newRequestCondition.length - 1][1],
+          ),
         )
       }
     }
@@ -151,61 +158,59 @@ export const CompoundRequestCondition = observer(
     }, [approx, requestCondition, statFuncStore])
 
     useEffect(() => {
-      setPreparedValue(
-        getResetType(requestCondition[requestCondition.length - 1][1]),
+      setPreparedScenarioName(
+        getScenarioName(requestCondition[requestCondition.length - 1][1]),
       )
     }, [requestCondition])
 
     return (
       <>
-        <>
-          <AprroxAndState approx={approx} setApprox={setApprox} />
+        <AprroxAndState approx={approx} setApprox={setApprox} />
 
-          <DividerHorizontal />
+        <DividerHorizontal />
 
-          <RequestConditions
-            problemGroups={problemGroups}
+        <RequestConditions
+          problemGroups={problemGroups}
+          requestCondition={requestCondition}
+          activeRequestIndex={activeRequestIndex}
+          onChangeScenario={onChangeScenario}
+          handleSetActiveRequestCondition={handleSetActiveRequestCondition}
+          onChangeRequestConditionNumber={onChangeRequestConditionNumber}
+        />
+
+        <div className="flex items-center justify-end w-full text-14">
+          <RequestsAmountControlButtons
             requestCondition={requestCondition}
             activeRequestIndex={activeRequestIndex}
-            onChangeScenario={onChangeScenario}
-            handleActiveRequest={handleActiveRequest}
-            handleRequestConditionNumber={handleRequestConditionNumber}
+            handleChangeRequestsAmount={handleChangeRequestsAmount}
+            onTouch={onTouch}
+          />
+        </div>
+
+        <DividerHorizontal />
+
+        <InheritanceModeSelect
+          handleSetPreparedScenario={handleSetPreparedScenario}
+          preparedScenarioName={preparedScenarioName}
+        />
+
+        <DividerHorizontal />
+
+        <div className="flex justify-between items-center mb-2 text-14">
+          <DisabledVariants
+            isFetching={isFetching}
+            status={status}
+            variantsValue={variantsValue}
+            variantsType={'True'}
           />
 
-          <div className="flex items-center justify-end w-full text-14">
-            <ControlButtons
-              requestCondition={requestCondition}
-              activeRequestIndex={activeRequestIndex}
-              handleRequestBlocksAmount={handleRequestBlocksAmount}
-              onTouch={onTouch}
-            />
-          </div>
-
-          <DividerHorizontal />
-
-          <InheritanceModeSelect
-            handleReset={handleSetPreparedValue}
-            resetValue={preparedValue}
+          <AllNotMods
+            groupSubKind={attributeSubKind}
+            isNotModeChecked={mode === ModeTypes.Not}
+            isNotModeDisabled={!variants?.length}
+            toggleNotMode={() => toggleMode(ModeTypes.Not)}
           />
-
-          <DividerHorizontal />
-
-          <div className="flex justify-between items-center mb-2 text-14">
-            <DisabledVariants
-              isFetching={isFetching}
-              status={status}
-              variantsValue={variantsValue}
-              variantsType={'True'}
-            />
-
-            <AllNotMods
-              groupSubKind={attributeSubKind}
-              isNotModeChecked={mode === ModeTypes.Not}
-              isNotModeDisabled={!variants?.length}
-              toggleNotMode={() => toggleMode(ModeTypes.Not)}
-            />
-          </div>
-        </>
+        </div>
 
         {controls &&
           controls({
