+++ src/components/conditions/custom-inheritance-mode/custom-inheritance-mode.interface.ts
@@ -25,7 +25,7 @@ export interface ICustomInheritanceModeConditionProps {
   statFuncStore: DtreeStatFuncStore | FilterStatFuncStore
 }
 export interface IHandleSetComplexScenarioProps {
-  preparedValue: string
+  preparedScenarioName: string
   problemGroups: string[]
   setScenario: (scenario: IScenario) => void
 }
