+++ src/pages/filter/common/attributes/base-functions.store.ts
@@ -4,6 +4,7 @@ import { ApproxNameTypes } from '@core/enum/approxNameTypes'
 import { DefaultProblemGroup } from '@core/enum/default-problem-group-enum'
 import { ModeTypes } from '@core/enum/mode-types-enum'
 import datasetStore from '@store/dataset/dataset'
+import { getApproxName } from '@components/conditions/utils/conditions.utils'
 import {
   ICompoundHetArgs,
   ICompoundRequestArgs,
@@ -15,7 +16,6 @@ import {
   TPropertyStatus,
   TRequestCondition,
 } from '@service-providers/common'
-import { getApproxName } from '@utils/getApproxName'
 import { getCurrentModeType } from '@utils/getCurrentModeType'
 import { BaseAttributeStore } from './base-attribute.store'
