+++ src/pages/filter/refiner/components/middle-column/panels/function-panel/components/compound-het/compound-het.tsx
@@ -7,12 +7,12 @@ import { FuncStepTypesEnum } from '@core/enum/func-step-types-enum'
 import { ModeTypes } from '@core/enum/mode-types-enum'
 import datasetStore from '@store/dataset/dataset'
 import filterStore from '@store/filter'
-import { AprroxAndState } from '@components/conditions/compound-request/components/approx-state'
+import { AprroxAndState } from '@components/conditions/components/approx-state'
+import { getApproxName } from '@components/conditions/utils/conditions.utils'
 import { AllNotMods } from '@pages/filter/dtree/components/query-builder/ui/all-not-mods'
 import { DisabledVariantsAmount } from '@pages/filter/dtree/components/query-builder/ui/disabled-variants-amount'
 import { ConditionJoinMode } from '@service-providers/common'
 import { ICompoundHetArgs } from '@service-providers/common/common.interface'
-import { getApproxName } from '@utils/getApproxName'
 import { getCurrentModeType } from '@utils/getCurrentModeType'
 import functionPanelStore from '../../function-panel.store'
 import { PanelButtons } from '../panelButtons'
