+++ src/pages/filter/refiner/components/middle-column/panels/function-panel/function-panel.store.ts
@@ -1,11 +1,7 @@
 import { makeAutoObservable } from 'mobx'
 
-import { FuncStepTypesEnum } from '@core/enum/func-step-types-enum'
 import filterStore from '@store/filter'
-import {
-  IFuncPropertyStatus,
-  TFuncCondition,
-} from '@service-providers/common/common.interface'
+import { TFuncCondition } from '@service-providers/common/common.interface'
 
 class FunctionPanelStore {
   constructor() {
@@ -15,16 +11,6 @@ class FunctionPanelStore {
   public get simpleVariants(): string[] {
     const { variants } = filterStore.statFuncData
     return variants
-  }
-
-  public get problemGroups(): string[] {
-    return (
-      (
-        filterStore.initialStat.getAttributeStatusByName(
-          FuncStepTypesEnum.CustomInheritanceMode,
-        ) as IFuncPropertyStatus | undefined
-      )?.family ?? []
-    )
   }
 
   public submitConditions(condition: TFuncCondition): void {
