+++ src/pages/filter/refiner/components/right-column/func-filter/components/compound-request-view.tsx
@@ -2,9 +2,9 @@ import { ReactElement } from 'react'
 import cn from 'classnames'
 
 import { approxOptions } from '@core/approxOptions'
-import functionPanelStore from '@pages/filter/refiner/components/middle-column/panels/function-panel/function-panel.store'
 import { getScenarioValue } from '@pages/filter/refiner/components/right-column/func-filter/utils/get-scenario-value'
 import { ICompoundRequestArgs } from '@service-providers/common/common.interface'
+import { refinerFunctionsStore } from '../../../attributes/refiner-functions.store'
 
 interface ICompoundRequestViewProps {
   isFilterActive: boolean
@@ -16,7 +16,7 @@ export const CompoundRequestView = ({
   isFilterActive,
   filterExpression,
 }: ICompoundRequestViewProps): ReactElement => {
-  const { problemGroups } = functionPanelStore
+  const { problemGroups } = refinerFunctionsStore
   const approx: string = filterExpression['approx'] || approxOptions[2]
   const state: string = filterExpression['state'] || 'current'
   const request = filterExpression['request']
