--- src/utils/function-panel/getFilteredRequestCondition.ts
@@ -1,9 +1,0 @@
-import { TRequestCondition } from '@service-providers/common'
-
-export const getFilteredRequestCondition = (
-  fullRequestCondition: TRequestCondition[],
-) => {
-  return fullRequestCondition.filter(
-    ([, reqCondition]) => Object.keys(reqCondition).length > 0,
-  )
-}
