--- src/utils/function-panel/getPureRequestString.ts
@@ -1,2 +1,0 @@
-export const getPureRequestString = (request: string): string =>
-  request.slice(8)
